:recycle: moved boxes and spheres to different files
remove collider cpp since it was basically the code for boxes and shperes but keeped the hpp with the interfact og a collider
