+++ src/physics/CMakeLists.txt
@@ -7,7 +7,10 @@ target_sources(StellarForgePhysics
         PUBLIC
         ${CMAKE_CURRENT_SOURCE_DIR}/Physics.hpp
         ${CMAKE_CURRENT_SOURCE_DIR}/Collider.hpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/Box.hpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/Sphere.hpp
         PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}/Physics.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/Collider.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/Box.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/Sphere.cpp
 )
