+++ src/physics/Physics.hpp
@@ -11,6 +11,8 @@
 #include <glm/vec3.hpp>
 #include <glm/gtc/quaternion.hpp>
 #include "Collider.hpp"
+#include "Box.hpp"
+#include "Sphere.hpp"
 
 /**
  * @typedef Vector3
